fix return value on pam_get_user failure
The return value of pam_sm_authenticate is supposed to be one of the
defined PAM_ return values. However, as a result of a broken
refactoring, we currently return -1 in case pam_get_user() fails. The
path this value takes is complex, but AFAICT it can bubble up to
pam_authenticate and possibly confuse the caller, though it's unlikely
that they'll mistake it for PAM_SUCCESS (which is numerically 0) so the
impact is probably limited.

Either way, fix it by returning a proper error code.